trading/pyproject.toml
[tool.poetry]
name = "crypto_ml"
authors = ["nerap"]
version = "0.1.0"
description = "Crypto trading ML system"

[tool.poetry.dependencies]
python = "^3.10"
torch = { url = "https://download.pytorch.org/whl/nightly/cpu/torch-2.2.0.dev20240101%2Bcpu-cp310-cp310-linux_x86_64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/nightly/cpu/torchvision-0.17.0.dev20240101%2Bcpu-cp310-cp310-linux_x86_64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/nightly/cpu/torchaudio-2.2.0.dev20240101%2Bcpu-cp310-cp310-linux_x86_64.whl" }
pandas = "^2.1.0"
numpy = "^2.2.1"
scikit-learn = "^1.3.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
plotly = "^5.18.0"
optuna = "^3.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.10.0"
isort = "^5.12.0"
mypy = "^1.6.0"
jupyter = "^1.0.0"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/nightly/cpu"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
